pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram-stc-bot"
version = "1.0.0"
description = "STC Charging Telegram Bot with Balance System"
requires-python = ">=3.9"
dependencies = [
    "aiogram>=3.13.0",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "python-decouple>=3.8",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "handlers*", "keyboards*", "middlewares*", "utils*", "config*"]
